tg/requirements.txt
aiogram==3.15.0
python-dotenv==1.0.0
aiofiles==24.1.0